Fix Link component conflicting with Angular router
This allows to create <a> tag without "href" property.
The necessity of this property comes from the fact that angular router is in conflict with react-router's Link component. To be more precise, Angular is [listening to all click events][1] and if the click is on an <a> tag with an href that [it recognizes as one of its own routes][2] then [it calls preventDefault() on the event][3]. The Link implementation [checks for default being prevented][4] and does nothing if it has.

[1]: https://github.com/angular/angular/blob/cf269d9ff43b913dbac4ba0ed6932c708ae58512/modules/angular1_router/src/ng_route_shim.js#L285
[2]: https://github.com/angular/angular/blob/cf269d9ff43b913dbac4ba0ed6932c708ae58512/modules/angular1_router/src/ng_route_shim.js#L294
[3]: https://github.com/angular/angular/blob/cf269d9ff43b913dbac4ba0ed6932c708ae58512/modules/angular1_router/src/ng_route_shim.js#L296
[4]: https://github.com/ReactTraining/react-router/blob/master/modules/Link.js#L78